+++ lib/src/extension/extension.dart
@@ -1,1 +1,3 @@
 export 'src/int_extension.dart';
+export 'src/iterable_extension.dart';
+export 'src/string_extension.dart';